changed some unsigned int to size_t for 64bit support
git-svn-id: http://svn.berlios.de/svnroot/repos/plow/trunk@51 fd8e78f2-7422-0410-845a-d0116f2ede31

+++ src/ID3v2Reader.cpp
@@ -85,7 +85,7 @@ void ID3v2Reader::read()
 
     std::string s;
     ID3v2::TextIdentificationFrame *tif;
-    unsigned int pos;
+    size_t pos;
 
     if(!(tag->frameListMap()["TRCK"]).isEmpty())
     {
